work around crasher bug in alsa-lib when setting samplerate
on my system (with an "nvidia CK804" onboard soundcard), i get immediate crashes
if i select the plughw device (in Pd this shows up as "NVidia CK808 (plug-in)")
with 44100Hz (the device only supports 48kHz)

with the help of the alsa-developpers i hunted the problem down;
it can be easily worked around if Pd was not using
snd_pcm_hw_params_set_rate_min() (without setting a rate_max()) but
snd_pcm_hw_params_set_rate_near() instead.

this patch does this, so no i can select all devices that show up in the media
menu without having to fear a crash.