use tagged LF for verbose() and friends
traditionally, verbose(), error(), bug(),... got their terminating LF by calling
endpost() since endpost() will output the LF 'untagged' (agnostic of the
log-level), this looks ugly when filtering-by-loglevel is applied on the
gui-side (the contents gets filtered but the the LFs bypass the filter) this
patch appens LF to the message and prints everything with the same loglevel.

also note that i the verbose-level offset is now set to '3' in relation to the
logpost-level (used to be '4').
this means that verbose(0, "foo"); will now already be visible when filtering
for 'debug' messages, which makes it (imo) much more useful.