parameterized paths to includes/libs
using relative paths for includes/libs directly is a bad idea, as it prevents
building out-of-the-sourcetree.
instead, we should use $top_srcdir (or $srcdir) and $top_builddir (or $builddir)
for the sourcetree resp the buildtree (which need not be the same).

this allows for easy testing whether the build-system is self-contained using
the `make distcheck` target.