provide sys_close(),... on all platforms
removing sys_close (as was done with commit:78b81aa3) will break binary
compatibility with externals compiled against Pd-0.43.
this patch re-introduces sys_close() and also provides stubs for the rest
of the open/close family.

while we are there, we also run the filename through sys_bashfilename
(which currently is a noop on non-w32 platforms)