don't connect to running pd-gui's if not needed
if we start multiple instances of Pd from the cmdline, we don't want to
connect to running instances anyhow.
since we already check, whether we should connect in parse_args{}, we don't
need to duplicate the check (in a buggy way!)

the current test, whether to look for other instances of pd-gui is the same
as the test, that tells pd-gui to connect to a running pd: if port>0 and host!=""

we use the full scriptname as the singleton key, so we don't have
sideeffects on various installations of Pd on the same system.