Use default searchpaths for stdlib/stdpath
the [declare] object allows to add search-paths and load libraries,
relative to the "Pd installation".
However, it does not honour the "standard" (immutable) search-paths
(as stored in `sys_staticpath`).

This patch will make '-stdpath' and '-stdlib' use all "standard" search-paths.

When using the "standard" search-paths, any leading "extra/" will be stripped.

This allows to use [declare -stdlib zexy], which will load the "zexy"
library if it can be found in one of ~/pd-externals, /usr/local/lib/pd-externals
/usr/local/lib/pd/extra/ (in case Pd is installed in /usr/local)