make [declare] have an immediate effect in an interactive session
adding [declare -path foo] now immediately adds "foo" to the search paths,
rather than having to save the abstraction and re-open it to have an effect.

this is only enabled in interactive-sessions (that is: not while a patch is
being loaded), to not interfere with the declare statements at the beginning
of a patch-file.