iterate over paths, and then over loaders
rather than iterating over the loaders, asking each to search for classes
wherever they like.

this allows us to override a class provided by an external with an abstraction,
simply by putting it into a path that is searched earlier.

for this to work, the loader-callback is extended to provide an additional
argument to the <path>.

if everything fails, we call the loader-callbacks with a NULL-path (telling
the loaders they should do how they please)