allow additional deken to override built-in
if somebody manually installed the deken-plugin, they probably want to use that
instead of the built-in, so the built-in should be loaded before.
this ensures that a manually installed plugin will shadow the built-in by
default.
by also providing a '::deken::version', the addon can check whether the built-in
is actually more up-to-date.