check whether a plugin-dispatcher is registered for a receiver
the original code will unconditionally iterate over
$::pd_connect::plugin_dispatch_receivers($receiver).
however, if $receiver was never registered for plugin-dispatching, this
will give an ugly backtrace.

this patch silently ignores plugin-dispatching for unknown receivers
(assuming that the abstraction finds a way to handle a
non-installed/non-responsive plugin).

an alternative would be to print a single-line warning (rather than a Tcl-backtrace).