updated iemgui's color-functions for loading/saving
we now accept symbols as colors, e.d. "#FF0000" is red.
we don't accept english colornames though (e.g. "blue"),
since internally the color is still handled as (int).

saving still uses the old lossy encoded-as-int-as-float representation,
in order to allow patches saved with Pd-0.47 be opened with Pd-0.46 or older.
once we reach 0.48, the default should be to store colors as symbols
(the old behaviour can be forced by setting the compat-mode to "47" or lower)