fixed loading of abstractions via absolute filename
class_new() uses some magic to *also* register the "long" path
(if that was given to instantiate the object), but class_addcreator()
lacked this.
The easiest way to get this behaviour for abstractions, is to use class_new()
as well.
however, this generates a new class for each abstraction (name), which we
will never use. to prevent a glaring memleak, we keep track of all those
classpointers via a linked list.