Patch from yarl:
pd is launching 1 more thread for each pd already running, due to:
(quote from http://jackaudio.org/api/types_8h.html)
JackNameNotUnique: The desired client name was not unique. With the JackUseExactName option this situation is fatal.
Otherwise, THE NAME WAS MODIFIED by appending a dash and a two-digit number in the range "-01" to "-99".
The jack_get_client_name() function will return the exact string that was used.
If the specified client_name plus these extra characters would be too long, the open fails instead.

I tried to keep the intended behavior but delegated renaming work to jack.