Update of expr-help.pd
Hi, I made a major overhaul in the documentation of expr. Basically, I kept all the original information, but reorganized things and added a LOT of missing information. Now we have every single function documented and separated into groups, that's the main addition!

I also added a version number and some useful examples/information on how to use [expr] with value/arrays and deal with "$0".