Add "~/Documents/Pd" as a Standard Path for MacOS
Hi, I Added a new standard path for mac os in "~/Documents/Pd", but kept "~/Library/Pd" for backwards compatibility. I've tested it, and it works great.

This deals with inconveniences in managing/installing externals since OS X Lion (10.7, released in 2011), when the current standard user folder (~/Library/Pd) was hidden by default. This makes it really hard managing externals there. For example, if you want to add it to the preference=>Path, you can't navigate it there, unless you know it is a hidden folder and how to make it visible to you. But making the folder visible might be not desired, so another option is knowing how to get there and make aliases. As someone who's teaching Pd workshops, I can tell you that's always hard and confusing to students!

Another option could be to access all standard folders from the Pd interface, such as from the "Path" tab in preferences. though this wouldn't require "hacks", it'd be a much more complicated fix, which would require a big interface overhaul. But the solution here in this Pull request makes it easier and it's a very simple addition.

This is not something unprecedented in Pd. For other reasons and issues specific to linux users, the last update (Pd 0.47-1) added a new standard path for Linux, namely: "~/.local/lib/pd/extra"

This update to Mac Os has also been discussed in the Pd list after the release of Pd 0.47-1. I realized It is more common to install such files as addons/plugins in somewhere like ~/Documents, this is where Max/MSP uses, for instance, to install externals.

Other have agreed with me in that discussion. Mark wrote:
"+1 on this... i just had a classroom of students navigate to ~/Library and create
~/Library/Pd .... its quite a pain to teach some of these basic skills. (...) 
BTW arduino and processing also both install libraries in: ~/Documents"
source: https://lists.puredata.info/pipermail/pd-list/2017-03/118277.html
And Dan wrote:
"I agree that something in the ~/Documents would be more standard than anything in ~/Library as that has been a hidden folder in Finder since OS X 10.7 or so as is not meant to be directly user-accessible. The preferences file, OTOH, should stay as that uses the same format & location that most macOS apps use."
source: https://lists.puredata.info/pipermail/pd-list/2017-03/118338.html
So, I hope I've made a good case that we'd benefit a lot if we'd also change this in the next release, in the same way it was last changed in linux for the 0.47-1 release.
thanks