function to check first creation argument against predefined symbols
this fixes an issue when you create certain [text] methods with a flag (or another predefined symbol) as the first creation argument which is then wrongly interpreted as the text buffer name.

consider: [text sequence -g] or [text search > 0].
the new function allows "-g" or ">" as the first creation argument, leaving the text buffer name empty.
this might break patches which use text buffers with the same name as those predefined symbols (which is quite unlikely, though).

another possibility would be to explicitly disallow those symbols as the first creation argument, posting an error message and asking the user for a dummy first argument.