properly reload cloned abstractions if one instance is saved
we don't want to remake the whole clone object when we save an instance; instead we keep the current instance and remake all other instances.
we need to check if the inlets/outlets have changed; if so, we just abort the reload and remake the whole clone object.
we also make sure that the saved instance is properly connected (outlet connections might get lost during editing)

+ move canvas_suspend_dsp()/canvas_resume_dsp() out of clone_setn()