Add menu item for connecting selected objects
Add a "Connect Selection" Edit menu item (shortcut "$accelerator-K")
that will heuristically attempt to connect selected objects.

The current implementation checks to see if there are only two
selected objects; if so, it connects the first outlet of the upper
object to the first inlet with a compatible type in the lower one.