add "delete" method
this deletes the current scalar and outputs the next one in the list (or sends a "bang" to the last outlet if it was the last one in the list).
we ensure that the gpointer stays valid so you can, for example, iterate over a glist and conditionally delete scalars.