adding resize flag to garray_read
I used fgetc() and counted newline characters to get file size. I based this on how Pd writes arrays to disk, so I am assuming that the input would be formatted in this way. I also used rewind(), otherwise calling -resize twice would give an error. Not sure about any of this, but it seems to work over here.