using 'open' instead of 'click' to open [text] window
The help file of [text] includes the messages 'click' and 'open', saying they can open and close windows.

But I don't believe it was the intention of Miller to document this, and if I'm not mistaken, someone else put that in the help file.

The thing is that a 'click' message is actually what you get by clicking on the object, and the same thing can be done to a [qlist] object, for instance - but that's not documented.

Since this has been documented and I'm using it. I think it's better to generate a more proper 'open' message name, which is just an 'alias' of 'click'.