don't draw [vu] outlets
This is one solution that came up in #632 and closes it.

Another solution to that issue would be to add a send symbol and a new extra argument. But the discussion led to this solution as probbaly easier and more sensible.

In practice, this hides the outlets without removing them for backwards compatibility concerns. The object keeps its outlets, but these are undocumented and maybe no one ever used them anyway as there doesn't seem to be much of a point in using them...

If this PR is not accepted I can work on the other solution (adding a send symbol).