Add "constant float" to [trigger] and allow "anythings" to be converted
This Pull Request also closes https://github.com/pure-data/pure-data/issues/583

This partially deals with the request https://github.com/pure-data/pure-data/issues/354 and takes care only of constant floats (not symbols). This is is also a simplification of https://github.com/pure-data/pure-data/pull/582 that implements constant values for floats and symbols. In this alternative, the the behaviour for symbols is exactly the same as before and this option seemed to encounter less resistance, so I thought I'd offer this option as well (see discussion in https://github.com/pure-data/pure-data/pull/582 for more details).