documentation updates
for adc~/dac~ I explicitly added the information that, by default, the loaded channels are 1 & 2 (it was just saying it was "stereo by default", but didn't specify which channels were loaded.

The other two files (acoustics/acoustic~) I didn't really change anything, just reorganized it a little with the spacings for presumably better looks.

I plan to do this kind of revision for all help patches now, here we go!