update fudiformat/fudiparse documentation
This one also had a smaller font size, but they also looked pretty harsh on the eye.

I added a little bit of information, saying what "fudi" is and stands for. I also added information about the -u flag in fudiparse and when the object and documentation was added/modified for both - as usual in all pd help files (in this case it was "updated for Pd 0.48.0)