hradio
where to start? This is also another help file with a font size of 10 that I'm changing to 12, but there's more going on.

There's an extra, old and redundanct help file here named 'hdial' that is not being used, so I'm just getting rid of it.

I also corrected an error that said "gui-vertical dial", but this is horizontal, and I think 'dial' is not to be used anymore, so I changed it to "gui horizontal radio".

I also reorganized and realigned things and changed some silly things, like having a float into a "$1" message then somewhere else. I got rid of that "$1" message because it was completely useless.

I guess this is because of this https://github.com/pure-data/pure-data/blob/fe7d40ebf75ad4ac31f67b52ae99485639f1ecbc/doc/1.manual/x5.htm#L986

but the thing is you can't create the [hdial] objects anymore, so I'm just letting it go...