Update numbox2-help.pd
changed font size, reorganized and realigned things.

Removed the [pd once] example as I don't think it's pertinent (like in the vsl/hsl examples)

Moreover, I realized it implies the object can be called as "numbox", but it can't. this is implied in the top left canvas "numbox=nbx". The actual creator is "my_numbox", so I fixed it and changed it to "my_numbox=nbx".