Update my_canvas-help.pd
I made it clearer what's the deal with setting the colors. The patch was using bioth 2nd and 3rd argument of the color message to set the 'label' color. So I removed the 2nd to only use the 3rd. I also made it clearer that the behaviour change was that this used to be the 2nd argument in esrlier versions of Pd, as it wasn't clear at all what changed!