Update pd-help.pd
Pd help seems to have been used as the help file of 'subpatches' (as in the 'pd' object). But this is not true anymore, as the canvas-help.pd file is now responsible to document subpatches.

Thus, I removed the implications that pd-help.pd was meant to be the documentation of [subpatch], and I just have [pd] included in the "see also" section.