revise control examples
This commit just reorganizes and realigns all the patchesa in the control example folder. Virtually no edits were made. All I ever did was make it a little bit more explicit that objects like [trigger] and [float] can be abbreviated in the example patches.

I the review part, I also just put the information that you can see all of Pd objects by right clicking on an empty spot and get help.

I remember I also told that in 17.PART3.midi.pd you could see all of pd's midi objects by getting the help of any of the objects in the example patch.

And that was it!