Revising and realigning all audio examples
So, this is where I was heading to all along. I always really thought this section needed some polishing, as some examples were not fitting in the screen, some tables and comments were overlapping with other elements in the screen, and it could just look better.

I know how messing with these files may seem bold, so I really tried to be least intrusive as possible and just do my regular realiging and reorganizing things spacially.

One exception is that in adsr.pd file I removed some connections and did include a send/receive as this was the only sensible way to me to avoid crossing wires over comments and the like...

Some examples will not look much significantly different. It was only very few examples (up to 5?) that I just felt I had to change/swap some of the elements (like, moving the text to another area of the patch so it would all fit better), the vast majority just kept the same structure. If that's an issue (I just cannot really think it could be), I'll go over the whole thing and specify which ones these are.

I was able to catch a few typos and little mistakes. One, for instance, is that E03.octave.divider.pd had a message for loading a so called "../../saucisse.wav" file that doesn't really exist, so I removed it. see https://github.com/pure-data/pure-data/blob/f9b0ad81387d4af3091857dc16852e221fe96310/doc/3.audio.examples/E03.octave.divider.pd#L114

Some of the abstractions used in the examples had comments saying they were loaded in examples that don't really exist anymore so I fixed that too.

Anyway, there are other issues I didn't really address yet, as I thouhgt it would be best to make separate commits for them. This is just a massive block of simple and non intrusive "fixes" that I think really improve the reading of patches.

I'll double checxk to see if I eventually corrupted something, but it's quite unlikely.

by the way, sorry for going totally anal about all of this... I just had to :/

cheers