Update openpanel-help.pd
This file has been edited and wqas a new conflict a had to deal with. So I redid the realignment but also fixed an error and mentioned this is being updated for 0.51 and not 0.50

I also went ahead and included an example of how to use [pdcontrol] to make openpanel open in the default directory.

I also changed the information that bang would open the 'current' directory as that wasn't accurate.