kbdnav in g_text.c
kbdnav stuff mainly responsible for creating new obj/msgbox/atoms
in the correct position and connecting them to the correct in/outlet.

Also there is kbdnav_iopos() which provides a single function to get
in/outlet positions.

Currently multiple functions of the pd core use copy/pasted
routines to deal with that. I intend to unify them to use this new
function later.