Check for multiple dollar signs in binbuf_restore
In the event that the first dollar sign we come across isn't related to a dollsym atom, we need to keep looking.
Otherwise, there will be inconsistencies between a patch's state upon saving, and its state after closing and re-opening.
Pd will save a backslash to the dollsym, but then fail to find it again upon re-opening.
The dollsym will be interpreted as a symbol, and the backslash will become visible in message boxes.