keep atoms i/olets functional
This keeps the inlet/outlet of atom boxes functional, like iemguis, when you set a send/receive name - as opposed to removing them.

This closes https://github.com/pure-data/pure-data/issues/1005

I reckon some invasive reorganization of the code needed to be made, I made my best to keep it not too wild, but a revision is needed as perhaps some other strategy could be better.