keep gatom i/olet functional with s/r names
This commit does not delete the inlet/outlet of gatom boxes when you insert a send/receive name. This make it consistent to the behavior of iemguis such as [nbx].

For this, I created a special atom_drawborder function that is called under gatom_vis and gatom_displace, this function calls a second glist_drawiofor_atoms function to draw the inlet/outlet.