bng-help reivision
I made yet a new revision of the bng-help file, adding yet more information. The main difference is that I'm now documenting a new way to set the colors of iemguis with hex notation. In this new version, I also included an example of the "old way" of setting colors.