make gatom flash when toggling
this is an improvement over the visual cue for gatom activity. Whem toggling in the number box, the object flashes only, like a message box - because when you do it, the object does not stay active so you can keep typing number values into it.

this is yet another commit that now fully fullfills and closes https://github.com/pure-data/pure-data/issues/1013