make atom width = 0 the same size as a setted width
this makes the size consistent between atom boxes with width = 0 to the soze of atom boxes with setted sizes.

The discrepancy was very very minimal. In order to make it fit, the boxes with setted width need to to be 2 pixels wider and pixel higher - this height is also consistent to the height of objects and other elements.

closes https://github.com/pure-data/pure-data/issues/1017