update iemguis
This is a major overhaul in the documentation of iemguis. The main motivation was to add the new undocumented way to set colors with hex notation, so this PR closes #1019 - there are still examples that show how to set the old way of setting colors.

I got carried away and just added detailed information on all parameters and added new examples. More notably, I added an example on how to use [vu] with env~ and slop~ (which closes #977 )

In this process I also encountered other documentation issues and fixed them all. Hence this PR also

- closes #1020
- closes #993
- closes #1003
- closes #1000

I also simplified the x_all_guis.pd file and removed the need for it to be called with arguments that were loaded inside it. It wasn't very clear what it was trying to illustrate with it, but now all the possible information is clearly described in the documentation anyway.