jack_out_max should really be the JACK blocksize (and not have a minimum of 64)
this allows us to use small JACK blocksizes (eg: 16 frames).
we still need to recompile Pd with a DEFDACBLKSIZE that is not bigger than
the desired JACK blocksize.

i have found no indication as to why the original code had a minimum of JACK_OUT_MAX (aka: 64).
it seems it was there from the very beginning (at least it was there in Pd-0.37.0
(somewhen in 2007), which i think was the version of Pd that introduced JACK support)

Closes: https://github.com/pure-data/pure-data/issues/1162