fix detection of locked samplerate/blocksize
the samplerate/blocksize is only fixed, if the *actual* audio-backend
prevents changing it.
if the user just requested a specific backend but couldn't get it (e.g.
because the audio server is not running), then we are free to assume
whatever sr/bs we want.

Follow-Up: https://github.com/pure-data/pure-data/pull/1287
Closes: https://github.com/pure-data/pure-data/issues/1283
https://github.com/pure-data/pure-data/issues/1283https://github.com/pure-data/pure-data/issues/1283https://github.com/pure-data/pure-data/issues/1283