Fix crash caused by incorrect JACK client name size limit
This removes the arbitrary `100` byte limit on JACK client names in
favour of using the limit specified by JACK itself.

From the [JACK docs][1]:

> `jack_client_name_size()`
>
> the maximum number of characters in a JACK client name including the
> final NULL character. This value is a constant.

If I `printf` the value locally, the constant turns out to be `128`.
This is the constant defined by the PipeWire libjack implementation.

I ran into this issue as I happen to have a default client name on my
device (Dell XPS 9310) that is `108` bytes long heh! This meant that
when selecting "jack" in the "Media" menu, pd would instantly crash.

See NixOS/nixpkgs#118835 for how I debugged and solved this.

After applying this patch, I can now use pure data with JACK
successfully.

[1]: https://jackaudio.org/api/group__ClientFunctions.html