Let the WindowManager decide where to put windows
Closes: https://github.com/pure-data/pure-data/issues/348
Closes: https://github.com/pure-data/pure-data/pull/1296
Related: https://github.com/pure-data/pure-data/pull/44
Related: https://github.com/pure-data/pure-data/pull/108

this PR lets the Window Manager (WM) choose the position of the PdWindow ("Pd console") and the initial position of a newly created patch window.
given that window managers have invested significant brainpower into how to best place windows for the user (and some of them even let the user configure how they want this window positioning to happen), i think we should rely on their powers rather than inventing our own ad-hoc solution for the problem (which some people - including myself - find irritating).

once a patch window gets saved, the window position is stored within the patch and respected once the patch gets re-opened.

the PR also improves accurate storing/restoring of window positions, regardless of the size of the window borders.
(the only exception is on Windows, where a slim window with a wrapped menu-bar gets wrongly positioned; this has been a known issue for some time and is not a regression; it seems there really is no general solution for this problem)

for those who really want a fixed position of the main PdWindow, a trivial GUI-plugin can do that job.

1. Don't force the pdwindow into the upper-left corner of the screen - instead let the WM choose

this most likely will put the PdWindow in the center.
on my system (Linux, multi-monitor) it puts the PdWindow in the center of
the currently active monitor; which is nice if the leftmost monitor is
occupied by something else and you actually work on a righthand monitor)

2. store window position rather than frame position

gets rid of all the whacky window offset calculation.
fixes moving windows for good.

as a side-effect, windows stored with older versions of Pd will have an offset
(at least on X11) to their prior-position.
however, the relative window position is kept intact,
and the offset is positive (so the window won't vanish off-screen)

3. let WM choose initial position of newly made window