fix font-metric calculation for font-extremes
if the smallest font cannot hold the requested font-metric,
the original code would return "0 0 0" (which is bogus).
instead we now return the smallest font (even if it is too big).

also, we start with smaller fonts (size=1 instead of size=6)

the 'fit_font_into_metrics' is refactored so that it no longer has side effects
but instead returns the found fontsizes.
that also means that the calls have to be adapted.


Closes: https://github.com/pure-data/pure-data/issues/1285