added GUIs to help intro
I created a new section in help-intro to include all GUI objects in Pd.

It only inserts x_all_guis so people can check it. I also included the message box in this x_all_guis file cause I consider it some sort of GUI since you can interact with it by clicking on it and it flashes to respond to your action. Also, it is a way to include the message box in the mix/list of all objects.