update tabread4~
like I did with canvas help, I added the example patches (succh as B15.onset) as files so users can more easily access them. The array seemed to be corrupted with many indexes being the same and it looked like we only had 3 indexes, so I also fixed that and changed the array name because if you open the examples or other help patcches, they all have the  same array name and Pd complains.