fixed "dologpost" not to send up "verbose" messages if Pd isn't set to verbose.  This is to hide an outstanding bug: when verbose is set, file- search-path printout is generated in a subthread by read/writesf~ and this sometimes generates garbled messages to the GUI which hangs it.
The bug is still there but now only should occur when "verbose" is turned on.

Meanwhile, I think the "PD_CRITICAL", etc., defines in m_pd.h should at the
very least be changed to reflect that they're about log levels, for instance,
LOG_CRITICAL".  But then, I don't see any instances of any of these except
VERBOSE and DEBUG.  Further, I don't think there should be any "DEBUG" messages
in Pd proper.  And anyway, most of the time when you want to have a debugging
messages you want to turn verbose off, so that it's easier to look through them.
So I think the whole thing should be rethought.

But for this release I guess it's enough to rename PD_VERBOSE, etc, so that
we at least don't have things like  "PD_NORMAL" in m_pd.h