improve documentation of time units for [delay], [metro] and [timer].
instead of referring to [delay] for more information, I made the information clear in all examples and also added more information on how things behave for each object. For instance, it was never clear to me that we could change the tempo of a [metro] object and this would affect the change in the middle of the beat.