Update makenote-help.pd
it has always annoyed me that the velocity value was "3.2", which is a very weird velocity value. I put a more proper value of '100' and added more information on how [makenote] can deal with any kind of number, even negative values, while actual MIDI messages needd to be in the range of 0-127!