rtext: allocate extra null byte for utf8 functions
rtext works on unterminated strings, but it calls utf8 functions which expect the input to be null-terminated.
this can lead to all kinds of weirdness and memory corruption.

solution: whenever we allocate/resize the content of x_buf, we make space for an extra 0 byte.
x_bufsize, however, still behaves like strlen(), i.e. doesn't include the 0 byte.