remove unnecessary casts when printing object addresses
casting pointers to 'long' causes warnings in 64-bit Windows builds (because 'long' is really 32-bit)

Note that this cast only appears in these three spots.
There are literally hundreds of places where we *don't* cast the pointer before passing it to ".x%lx"
Also, strnpointerid() is only used in two place (in s_print.c). We really need a general solution for the whole code base (which doesn't have to be a function!), so I went ahead and removed this function.