remove setctl abstraction from the help files
this very very very simple abstraction was just used in one of the examples of the [slop~] help file. I removed it so we have less noise and not that necessary files (hence, to clean things up).

The necessary changes were made to the example and things ended up being cleaner in the patch as well, which I also think contributes to clarity in the patch.