minor cosmetic changes, standardized "Pd version"
I made is standard that the message is "updated for Pd version x.x". There mentionings to "release" and I also thought the standardd was to suppress "version" and made changes like that. It's all "Pd version" now.

Made also numerous cosmetic and minor changes such as alignments and moving things around a tiny bit.