added output~ to help-intro.pd and edited release notes
since we now have output~ in 'extra' I added it to the object list under the 'extra' category. I also mentioned this in the release notes. Ressi's collaboration were listed twice so I removed the redudancy. I also added how I did lots of work on the help files :) hope it's cool to brag about it :/