Update B15.tabread4~-onset.pd
I'm using the fast-forward message to Pd here to generate the table cause it's a very good use case. Also, this file now is called inside the help file of tabread4~ and it would turn the DSP audio on when opening it. And we have a strict standard to NOT EVER let audio turn on when opening help files now.