rename 'gatom-help' to 'gui-boxes'
In order to provide a more meaningful help file name (and again thinking ahead for a next step online reference), I'm choosing another name instead of 'gatom', which is just an internal object name not given or explained ANYWHERE besides the source code.

I also made it clear what are the names of each object in the top of the help file.