add reference to snapshot~ / threshold~ and trace, adding a bit more information to threshold~ and trace.
not much changes in the help files, added a bit more information for clarity in the right inlet of [threshold~].

The trace object has been edited so it is more in the standard of usual help files in Pd, with some description of the object coming first as a 'header' before giving an example. The example has also beem modified to offer more tracing options.