document pointers in [route] and better clarify special list cases
This commit document how [route] also deals with pointers and closes https://github.com/pure-data/pddp/issues/93

It also tries to be less ambiguous about handling special list cases such as 'list' becoming a bang. It wasn't clear that this operation was intentionally taken care by the object and I got misled that "Pd" would magically do this on its own whenever a message is sent to whatever destination. Most of this is discussed in https://github.com/pure-data/pddp/issues/94 and I still want to make this clearer and more in depth in Pd's manual.

The same issue was better addressed in other help files, such as [print] and [trigger]. Minor changes were made to [list] in [pd about-lists] so things were a bit clearer and I also added in that subpatch to "see also help file of message boxes" as it does explain a lot of what is going on there. [list trim] explained complicated selector handling behaviour that I think it shouldn't be explained and that it should be handled differently anyway, see https://github.com/pure-data/pure-data/issues/1679