still clarifying weird list cases (0/1 element list) and better mentioning the concept of atoms
We did get lots of things wrong when discussing an update of [route] and [trigger]. This is rather complicated, I started working on this on the last commit but more was needed.

As it is, I'm now documenting things correctly and properly to the current state. I have a PR and discussion on changing things a bit and will revise the documentation if changes are made.

I better documented the special case of a list box and an empty list input. I also worked a little bit in clarifying the concept of an 'atom' in the help file of GUI boxes and the control examples with minor revisions.