reverting it back changes from the last two commits about 'special case lists'
I have made things simpler according to discussions regarding my previous commits. I agree it was too much information that was not relevant to the user. I just kept this information on the [print] object as it felt relevant to me. I also think we should still document this also in the manual.