improve mute logic for output~
I'm now offering a much more sophisticated logic for muting [output~]. The bang label now also shows the internal state and goes to 'unmute' if you can unmute it.

Now the mute bang turns the volume off only if it's nonzero. It restores to the last setting when you click back on it to unmute, but if you raise the volume back with the number box, then it also unmutes.

Not plainly a 'documentation' issue but I did include this new object as part of this branch since it is used so much in the documentation.

Other changes in other files are just cosmetics.